fix: switch to containerd
The cri-o package for NixOS doesn't allow you to override the CNI bin
directory, which is a problem considering cilium's binary isn't in the
nix store. I was going to rewrite it but then I was advised by an OCF
alumni that cri-o is a world of pain, so we're switching to containerd.