revert "feat: add OCF root certificate"
Turns out, adding the certificates to the host doesn't actually affect
containers. This is obvious in hindsight (you need to modify a host file
system path, which is affected by overlayfs), but for some reason I had
it in my head that I'd need to add it on the host. Not adding it on the
host is likely good for security anyway.